Fixed a few bugs, better resiliency for downloads
For some reason, I was closing the connection to the database before the
DatabaseWorker was finished doing, which is pretty dumb. I also did a
join() on the StatsProcessor before all other threads were stopped,
causing a deadlock.

Moved the USED_TO_DB dict to chrome_db.py because it just made more
sense for it to be there.

Ran into a similar ConnectionError I got when trying to verify an
extensions availability (util.py) when downloading the actual CRX. Added
similar code that waits for a couple seconds and then retries for a max
of 5 retries.

Reduced a few common logging messages to DEBUG level.