Added all selenium/autotest/telemetry dependencies
This includes symlinks from the src directory to the deps directory and
an updated requirements.txt file.

The files I used may be out of date, but it was difficult to find
documentation or links to them. I actually copied some of the libraries
from the Chromium project (not the Chromium OS project).

Most of the libraries were written for Python 2 and had to have some
work done to fit this project's needs for Python 3 as well as the
project directory structure. It is likely that some code incompatible
with Python 3 still exists and will need to be "massaged" into
compliance.