First (mostly) working versin of async crawler
Seems there's still an issue with the program terminating correctly.
Most likely there's a condition that a task is never marked as complete,
so when join()-ing on the queue, it blocks indefinitely. Finally
terminated the first real test run after it sat there for over a week
making no progress.