Working version :-)
Now, dbling will come up on a fresh machine (16.04) with no packages
installed manually (was so close before!). This includes one interesting
dependency of Ansible on a 2.X version of Python. Now, by specifying the
interpreter explicitly, this isn't an issue.

Created a hosts file, which will make it more portable to run Ansible
without having to edit the system-wide file at /etc/ansible/hosts. To
use this, call Ansible with the flag `-i hosts`.

Included some improved checking to the playbook as to whether it is
actually necessary to restart the celery daemons based on how previous
playbook commands went (changed or not). This should allow for adding
new workers to a pool while not disrupting any workers already running.

Improved logging when interacting with the Chrome Web Store (in
crawl/webstore_iface.py).

If the list download fails, it will now always send an email to the
admin.