Fixes #6 by disabling StrictHostKeyChecking
It turns out that the problem described in #6 was being caused by
failing host key validation. It wasn't clear that this was the issue at
first because the `crawler-sshfs` service is configured to restart on
any failure. This was wiping out the original error message and
replacing it with one about attempting to restart the service too
quickly after stopping it.

On a fresh worker VM created with the Ansible playbook, I temporarily
removed the restart setting, which showed the host key verification
error, which simply means that the fingerprint of `dbling-master` is
unknown. This commit fixes this issue by disabling StrictHostKeyChecking
in the sshfs configuration and emptying the known hosts file on every
run of the playbook. This will also help when a VM is rebuilt and has
the same IP as a previous worker or master.