Name and version info from manifest
Crawler now reads the name and version of an extension from the manifest
included in the CRX. This information is read immediately after the CRX
is extracted, then the info is added to the database.

The database structure has been altered to accommodate this new
information. In the process of doing this, I discovered the length
requirement for both the name and version fields, eliminating the need
to guess as to the upper bound of how long these strings may be.