Directory size calculation test changes
These changes are very closely connected to c6636d3 and cb8001e, all of
which were in preparation for an experiment testing whether a change in
how we predict the size of a directory on an encrypted filesystem would
improve the accuracy of a centroid in matching the correct extension.

The results of the experiment were not favorable. It's possible these
changes helped things, but at this point it's difficult to say how much.
One thing is clear, the predicted size of directories were very
frequently no bigger than its size when not yet encrypted.

Other improvements to the code include:

- Correction of some import problems by moving a few more things to
  `const.py` (which imports no other files from the project)
- Allowing the DblingGraph class to be copied such that the resulting
  object is also a DblingGraph with all the properties copied to the new
  object.
- A utility function for calculating the byte length of `str` and
  `bytes` objects.
- Optimization of imports in several files.
- Easier switching between MERL and plain output when running the
  profiler.
- Auto-closing of the connection to the DB at exit from the `merl`
  module.