Encrypted files don't need fancy calculations
When an encrypted file is being processed, we don't need to (and
shouldn't) do any calculations to determine its size. The calculations
we have to do are only when working with unencrypted files to predict
their size after encryption. This commit makes changes that align the
code with this principle.

Also, removed code that included `ctime` in some calculations leading up
to the centroid calculation. The `USED_FIELDS` constant prevented it
from actually being included in the centroid, but these changes remove a
couple unnecessary calculations, pulled from 2f667824.