Encrypted temporary directories!
The code now uses a pair of temporary directories in conjunction with
eCryptfs to provide an "upper" and "lower" filesystem just for
unpacking. What this means is that (1) the unpacking still happens
completely in memory, but (2) we're now able to mimic the properties of
the files when they're encrypted, because we point the profiler at the
encrypted directory! :-)

Also, removed some old, unnecessary `__future__` imports and the old
crontab file.