Completely changed how dbling summarizes jobs
Added a new node just for summarizing jobs that binds to a queue just
for that purpose. There has to be only one node with this role for
things to work, though, because it keeps a running tally of statistics
in a local file. That file can only be accessed by one piece of code at
at time, locked by using memcached to store a key that indicates when a
process is using it.

The summarizing node is necessary because the previous setup of putting
all the analysis jobs in a single chord wasn't scalable. Before the
entire job could complete (this was a re-profiling job that involved
over 200K extensions), RabbitMQ ran out of memory with. This doesn't
necessarily stop everything, but it keeps any workers from producing any
new messages. The only way out of this situation is to have some worker
that only *consumes* jobs (messages). In other words, I needed to add
something that could release some of the memory by finalizing some jobs
without adding any new messages.

So, I had to write a few new utilities to accomplish all of this:
- `chunkify`: Splits an iterable into smaller iterables. This is what
  helps me break the list of extensions into smaller lists. `chunkify`
  is smart about generators, though, and won't force them to expand to
  their full length when processing them.
- `acquire_lock`: Handles interfacing with memcached to get/delete the
  lock. I wrote this as a context manager so it can be used in a `with`
  clause. It can automatically wait for some specified number of seconds
  if the lock isn't available.

I also pulled out some of the Ansible tasks into a file to make it
easier to reuse them, which made it so the section defining the new
summarizing node is about 4 lines long.