First attempt at beacon extension prototype
Successfully retrieves the user's email address, but not the MAC
address. Also, there needs to be a mechanism that caches times in case
the device goes offline.

The manifest file already has the permissions necessary to get the IP
address. The IPv6 address can be used to derive the MAC address. See
https://stackoverflow.com/a/38994590 for more info on how to do this.